Переосмысливает рефакторинг согласно данным комментариям: добавляет фабрику объектов и конфигов, удаляет лишние отдельные файлы сцен и импорты объектов (теперь создаются внутри фабрики)
