+++ source/js/modules/3d/objects/lantern.js
@@ -1,10 +1,11 @@
 import * as THREE from "three";
 
 class Lantern extends THREE.Group {
-  constructor(materialsFactory, options) {
+  constructor(materialsFactory, config) {
     super();
     this.materialsFactory = materialsFactory;
-    this.options = options;
+    this.options = config.options;
+    this.materials = config.options.materials;
     this.lampBottomPosition = null;
     this.postBottomPosition = null;
     this.constructChildren();
@@ -26,7 +27,7 @@ class Lantern extends THREE.Group {
     // радиус окружности, описанной около квадрата, равняется длине его стороны, умноженной на квадратный корень из двух и деленной на два
     const radiusLampTop = (lamp.top.widthTop * Math.sqrt(2)) / 2;
     const radiusLampBottom = (lamp.top.widthBottom * Math.sqrt(2)) / 2;
-    const materialLampTop = this.materialsFactory.get(lamp.top.material);
+    const materialLampTop = this.materialsFactory.get(this.materials.lamp.top);
 
     let top = new THREE.Mesh(
         new THREE.CylinderGeometry(
@@ -45,7 +46,7 @@ class Lantern extends THREE.Group {
     // аналогично берем геометрию цилиндра и задаем ему 4 стороны, рассчитываем радиусы оснований
     const radiusPlafonTop = (lamp.plafon.widthTop * Math.sqrt(2)) / 2;
     const radiusPlafonBottom = (lamp.plafon.widthBottom * Math.sqrt(2)) / 2;
-    const materialLampPlafon = this.materialsFactory.get(lamp.plafon.material);
+    const materialLampPlafon = this.materialsFactory.get(this.materials.lamp.plafon);
 
     let plafon = new THREE.Mesh(
         new THREE.CylinderGeometry(
@@ -61,7 +62,7 @@ class Lantern extends THREE.Group {
     plafon.rotation.y = Math.PI / 4;
 
     // основание лампы представляет собой параллелепипед
-    const materialLampBase = this.materialsFactory.get(lamp.base.material);
+    const materialLampBase = this.materialsFactory.get(this.materials.lamp.base);
 
     let base = new THREE.Mesh(
         new THREE.BoxGeometry(
@@ -88,7 +89,7 @@ class Lantern extends THREE.Group {
   // добавление столба
   addPost() {
     const {post} = this.options;
-    const materialPost = this.materialsFactory.get(post.material);
+    const materialPost = this.materialsFactory.get(this.materials.post);
 
     let postMesh = new THREE.Mesh(
         new THREE.CylinderGeometry(post.radius, post.radius, post.height),
@@ -103,8 +104,8 @@ class Lantern extends THREE.Group {
   // добавление основания фонаря
   addBase() {
     const {base} = this.options;
-    const materialBaseTop = this.materialsFactory.get(base.top.material);
-    const materialBaseBottom = this.materialsFactory.get(base.bottom.material);
+    const materialBaseTop = this.materialsFactory.get(this.materials.base.top);
+    const materialBaseBottom = this.materialsFactory.get(this.materials.base.bottom);
 
     let baseTop = new THREE.Mesh(
         new THREE.SphereGeometry(base.top.radius),
