+++ source/js/script.js
@@ -29,3 +29,4 @@ social();
 
 const fullPageScroll = new FullPageScroll(scene3D);
 fullPageScroll.init();
+
